+++ src/zig_codeblocks/styling.py
@@ -1,4 +1,6 @@
-from dataclasses import KW_ONLY, dataclass
+from __future__ import annotations
+
+from dataclasses import KW_ONLY, dataclass, replace
 from enum import Enum
 from itertools import compress
 
@@ -17,7 +19,11 @@ class Reset(Enum):
 
 
 class Color(Enum):
-    # Names adjusted to match Discord's style
+    """
+    An enumeration of 3-bit ANSI colors.
+    Some names were adjusted to match Discord's style.
+    """
+
     GRAY = "30"
     RED = "31"
     GREEN = "32"
@@ -30,6 +36,12 @@ class Color(Enum):
 
 @dataclass(slots=True, frozen=True)
 class Style:
+    """
+    A style for syntax highlighting.
+    Takes a `Color` and can optionally be bold and/or underlined.
+    Produces an SGR sequence when converted to a string.
+    """
+
     color: Color
     _: KW_ONLY
     bold: bool = False
@@ -38,3 +50,22 @@ class Style:
     def __str__(self) -> str:
         modifiers = compress(("1", "4"), (self.bold, self.underline))
         return _to_sgr(self.color.value, *modifiers)
+
+
+@dataclass(slots=True, kw_only=True)
+class Theme:
+    """A theme for syntax highlighting Zig code."""
+
+    builtin_identifiers: Style | None = None
+    calls: Style | None = None
+    comments: Style | None = None
+    identifiers: Style | None = None
+    keywords: Style | None = None
+    numeric: Style | None = None
+    strings: Style | None = None
+    primitive_values: Style | None = None
+    types: Style | None = None
+
+    def copy(self) -> Theme:
+        """Create a copy of the `Theme`."""
+        return replace(self)