+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.0] - 2025-02-09
+
+### Added
+- Theming support
+- Improved reset code insertion
+- Improved type highlighting
+- Primitive value highlighting (i.e. `true`, `false`, `null`, `undefined`)
+
+### Changed
+- Swapped keyword and type colors in the default theme
+
 ## [v0.1.2] - 2025-02-09
 
 ### Added
@@ -30,3 +41,4 @@ Initial release 🎉
 [v0.1.0]: https://github.com/trag1c/zig-codeblocks/releases/tag/v0.1.0
 [v0.1.1]: https://github.com/trag1c/ixia/compare/v0.1.0...v0.1.1
 [v0.1.2]: https://github.com/trag1c/ixia/compare/v0.1.1...v0.1.2
+[v0.2.0]: https://github.com/trag1c/ixia/compare/v0.1.2...v0.2.0